Add files via upload
Hi I used a spring.io rest tutorial https://spring.io/guides/tutorials/rest/ . I did the the first half of the tutorial step by step. This tutorial is helping me gain the knowledge on how to make a functioning REST Api which I heavily struggled with at the beginning of this semester. I will be completing this tutorial as my third and final api project.